infrastructure/testdriver/actions/dragAndDrop.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>TestDriver actions: drag and drop</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-actions.js"></script>
<script src="/resources/testdriver-vendor.js"></script>

<style>
div { padding:0px; margin: 0px; }
#trackPointer { position: fixed; }
#resultContainer { width: 600px; height: 60px; }
.area { width: 200px; height: 100px; background-color: #ccc; }
.block { width: 5px; height: 5px; border: solid 1px red; }
.box { display: flex;}

</style>
<body>
  <div id="trackPointer" class="block"></div>
  <div>
    <h2>draggable</h2>
    <div class=box>
      <div id=draggable draggable="true" class="area"></div>&nbsp;
      <div id=droppable dropzone="true" class="area"></div>
    </div>
  </div>
  <div id="resultContainer">
    <h2>Events</h2>
    <div id="events"></div>
  </div>
</body>
<script>
let event_type = [];
let event_id = [];

var els = {};
var allEvents = { events: [] };
function displayMessage(message) {
    document.getElementById("events").innerHTML = "<p>" + message + "</p>";
}

function appendMessage(message) {
    document.getElementById("events").innerHTML += "<p>" + message + "</p>";
}

function recordPointerEvent(event) {
  event_type.push(event.type);
  allEvents.events.push({
    "type": event.type,
    "button": event.button,
    "buttons": event.buttons,
    "pageX": event.pageX,
    "pageY": event.pageY,
    "ctrlKey": event.ctrlKey,
    "metaKey": event.metaKey,
    "altKey": event.altKey,
    "shiftKey": event.shiftKey,
    "target": event.target.id
  });
  appendMessage(event.type + " " +
      "button: " + event.button + ", " +
      "pageX: " + event.pageX + ", " +
      "pageY: " + event.pageY + ", " +
      "button: " + event.button + ", " +
      "buttons: " + event.buttons + ", " +
      "ctrlKey: " + event.ctrlKey + ", " +
      "altKey: " + event.altKey + ", " +
      "metaKey: " + event.metaKey + ", " +
      "shiftKey: " + event.shiftKey + ", " +
      "target id: " + event.target.id);
}

function recordFirstPointerMove(event) {
  recordPointerEvent(event);
  window.removeEventListener("mousemove", recordFirstPointerMove);
}

function resetEvents() {
  allEvents.events.length = 0;
  displayMessage("");
}

async_test(t => {
  var draggable = document.getElementById("draggable");
  draggable.addEventListener("dragstart", recordPointerEvent);
  draggable.addEventListener("dragenter", recordPointerEvent);
  draggable.addEventListener("dragend", recordPointerEvent);
  draggable.addEventListener("dragleave", recordPointerEvent);
  draggable.addEventListener("dragover", recordPointerEvent);
  draggable.addEventListener("click", recordPointerEvent);

  var droppable = document.getElementById("droppable");
  droppable.addEventListener("drop", recordPointerEvent);

  let actions = new test_driver.Actions()
      .pointerMove(0, 0, {origin: draggable})
      .pointerDown()
      .pointerMove(5, 0, {origin: draggable})
      .pointerMove(10, 0, {origin: draggable})
      .pointerMove(15, 0, {origin: draggable})
      .pointerMove(20, 0, {origin: draggable})
      .pointerMove(-10, 0, {origin: droppable})
      .pointerMove(0, 0, {origin: droppable})
      .pointerMove(5, 0, {origin: droppable})
      .pointerMove(10, 0, {origin: droppable})
      .pointerMove(15, 0, {origin: droppable})
      .pointerUp();

  actions.send()
    .then(t.step_func_done(() => {assert_array_equals(event_type, [
      "dragstart", "dragenter", "dragover", "dragover", "dragover",
      "dragleave", "dragend"
      ]);}))
    .catch(e => t.step_func(() => assert_unreached("Actions sequence failed " + e)));
});
</script>
